requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2
djangorestframework==3.14.0
idna==3.10
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
